LoadActions fully refactored on new httpGet
