+++ client/app/js/http/fetchers.js
@@ -1,25 +1,32 @@
 import httpGet from './utils/httpGet';
 
-export function getToken(email, password) {
+export function fetchToken(email, password) {
   return httpGet({ url: '/api/token', auth: { u: email, p: password } });
 }
 
-export function getCurrentUser(token) {
+export function fetchCurrentUser(token) {
   return httpGet({ url: '/api/current_user', token });
 }
 
-export function getUserLeagues(userId, token) {
+export function fetchUserLeagues(userId, token) {
   return httpGet({ url: `/api/users/${userId}/fantasy_leagues/`, token });
 }
 
-export function getLeagueFantasyTeams(fantasyLeagueId, token) {
+export function fetchLeagueFantasyTeams(fantasyLeagueId, token) {
   return httpGet({
     url: `/dev_api/league/${fantasyLeagueId}/fantasy_teams/`,
     token
   });
 }
 
-export function getLeagueFantasyPlayers(fantasyLeagueId, token) {
+export function fetchLeagueFootballPlayers(fantasyLeagueId, token) {
+  return httpGet({
+    url: `/dev_api/league/${fantasyLeagueId}/football_players/`,
+    token
+  });
+}
+
+export function fetchLeagueFantasyPlayers(fantasyLeagueId, token) {
   return httpGet({
     url: `/dev_api/fantasy_leagues/${fantasyLeagueId}/users/`,
     token
